+++ composer.json
@@ -9,8 +9,13 @@
     ],
     "autoload": {
         "psr-4": {
-            "ByJG\\AnyDataset\\": "src/",
-            "AnyDataSet\\Tests\\": "tests/"
+            "ByJG\\AnyDataset\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Tests\\AnyDataset\\": "tests/",
+            "TestsDb\\AnyDataset\\": "testsdb/"
         }
     },
     "prefer-stable": true,
